+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
 =====
 
 * Add ``users/validate`` endpoint for bulk validation of user identifiers (usernames or emails).
+* Add org-wide support to migration commands for forward and backward migration of course authoring permissions.
 
 1.5.0 - 2026-04-09
 ******************